Initial conversion of JSS to SASS for reusable components
