+++ assets/components/TemporaryDrawer.js
@@ -2,15 +2,11 @@ import React from "react";
 import PropTypes from "prop-types";
 import Drawer from "@material-ui/core/Drawer";
 
-const TemporaryDrawer = (props) => {
-  const { drawerOpen, toggleDrawer, children } = props;
-
-  return (
-    <Drawer open={drawerOpen} onClose={toggleDrawer}>
-      {children}
-    </Drawer>
-  );
-};
+const TemporaryDrawer = ({ drawerOpen, toggleDrawer, children }) => (
+  <Drawer open={drawerOpen} onClose={toggleDrawer}>
+    {children}
+  </Drawer>
+);
 
 TemporaryDrawer.propTypes = {
   drawerOpen: PropTypes.bool.isRequired,
